mod fuzzy_finder
