+++ src/fuzzy_finder.rs
@@ -1,0 +1,1 @@
+pub mod fuzzy_finder;
